Working Engine
This commit saves the current state of the system generator engine for
doing the serial FIR filter, the decimation, and the template match. As
it is saved there is a sin wave with noise and a voltage offset stored
in the memory. A pure sin wave in the template memory, and the filter
memory is preloaded each run with coefficients generated and output from
the FDATool.
There is no sample timing, in that it is waiting for the
'new_sample_available' signal which has been tied high, so it goes AFAP
not waiting for the 50uSec between sample.